// frontend/src/apiservice/model/refuel.ts
/**
 * Vehicle database API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Entity } from './entity';
import { RefuelData } from './refuelData';
import { RefuelAllOf } from './refuelAllOf';


export interface Refuel {
    _id: string;
    createdAt: string;
    updatedAt: string;
    __v?: number;
    odometerValue: number;
    fuelAmount: number;
    price: number;
    refuelDate: string;
    notes: string | null;
    vehicleId: string;
    userId: string;
}